Enforce the constant fn in the obs/function space
This commit adds the required homogenous change of coordinates such that the OLS solution provides the optimal bias vector, which can be interpreted of the coefficient of the projection of each function in the observable space to the constant function.
This approach allow us to keep the transfer operator always free of the trivial 1 eigenvalue.